Fix some appearance issues (#1398)
* Fix property spelling in Sidebar component

* Fix typo in todos partition name

* Use the FerdiDev setting.json in dev mode

The app.setPath calls are still a mess and need to be fixed, but this at
least allows me to experiment locally.

* Make @meetfranz/theme follow the accent color

We inject the value of the accent color into the themes, so the buttons
show the appropriate color. This allows removing some existing css
hacks.

In order to do this, the themes had to be turned into functions that
take the brand color (accent color) as an argument instead of
hard-coding it.

The css-based accent color for legacy components is left unchanged.

* Remove superfluous scrollbars

While overflow: scroll doesn't look ugly on macs, because the system
hides scrollbars by default (except on hower), on Linux and Windows, the
disabled scrollbars appearing by default look ugly.

We set overflow-y: auto instead to display a scrollbar only when needed.

* Simplify theme generation in @meetfranz/themes

* Remove default accent color code duplication

* Apply accent color to styles not in themeInfo.json

Some colors are darkened/lightened versions of the primary brand color,
so they are not caught by build-theme-info. We create these styles with
`color` manually.

* Inset shadow on sidebar

This creates a more flat look that projecting a shadow over the service,
and avoid compositing a shadow over the webview.

* Apply accent color to vertical style tab bar

* Workspace drawer styling in vertical mode

* Fix sidebar layout

In vertical mode, the top of the service was clipped by the sidebar

Also removes unnecessary whitespace in sidebar

* Fix 1px wide line on right when todos are hidden

* Fix window menu

When building the menu bar, the File / app menu is added later, so we
must be careful about indexing the menu bar.

* Update locale files

* Report the default accent color accurately

* Fix inverted button with dark theme

* Fix crash with invalid accent color

Fall back to the default accent color instead

* Darker secondary buttons in dark mode

* Vertical workspace drawer style outside debug mode

In order to reliably apply the style, we need to add a class name to the
drawer manually, as component names only appear in class names in debug
mode.