Electron 13 support on Linux (#1447)
* Bump dbus-next to 0.9.2

Older dbus-next seems to have problems with building with Electron 13
ABI. Also upgrades nan (a transitive dependency of dbus-next) to 2.14.2.

* Explcitly specify Electron ABI 89

electron-rebuild seems to have problems with finding the correct ABI for
Electron 13, despite node-abi being updated to the latest version and
containing an entry for Electron 13 in abi-registry.json. This is
probably caused by Electron 13 still being marked as a future target in
abi-registry.json.

This commit pins the ABI version to 89 for Electron 13 and should be
reverted once node-abi marks Electron 13 as released.