fix: accent color customization regression (#1963)
PR #1959 removed the build-theme-info script, which was broken, along
with the themeInfo.json that we used to dynamically build CSS rules for
accent color customization. This resulted in the accent color not
getting applied to most of the UI.

Instead of trying to restore the build-theme-info functionality, this
commit takes a much simpler approach by statically adding the required
CSS rules to src/features/appearance/index.ts.

Since these rules are no longer automatically generated, we need to
update them once some CSS styling is added with the accent color.

A longer term fix would be to use CSS variables instead, but that
approach would not be supported by @meetfranz/theme.