refactor: preload-safe-debug shim
In https://github.com/ferdium/ferdium-app/pull/23 we removed usages of
the debug package due to an electron bug.

This patch aims to restore some debug functionality by introducing a
shim. The shim detect whether if it is being introduced in a preload
script where the electron but would be triggered, and falls back to a
simple replacement for debug. However, in the main and renderer
processes, where a preload script is not being used, we still get full
debug functionality.

In this way, a module can be used both in a preload script and outside
of it, while still preserving debug functionality whenever possible.

Signed-off-by: Kristóf Marussy <kristof@marussy.com>