Add new token request (#1384)
* Add new token request

This exists so that the user generates a new JWT token with the new adonisjs update on the server (that is now stored - hashed and using the adonis5-jwt package - which previously wasn't). This logic enhances security as we can delete the tokens on the jwt_tokens database in order to log-off users in case the APP_KEY is compromised (if we are hacked, for instance). After some time (maybe months after the migration) we can delete the old code on the server side that handles the deprecated JWT tokens (and possibly change the APP_KEY used to sign the old certificates - we can now probably use certs to sign the JWT - to enhance security as well).

* Update src/api/server/ServerApi.ts

Co-authored-by: MCMXC <16797721+mcmxcdev@users.noreply.github.com>

* Update src/stores/UserStore.ts

Co-authored-by: MCMXC <16797721+mcmxcdev@users.noreply.github.com>

---------

Co-authored-by: MCMXC <16797721+mcmxcdev@users.noreply.github.com>