Bring in missing python (system dependency) (#1639)
* Introduce python as a system dependency (needed by node-gyp to compile native modules)
* Introduce node-pre-gyp as a dependency
* Upgrade sqlite3 to 5.1.7 to test the above

Note: Currently this still depends on the availability of prebuilt binaries from the prebuild library. Which means that once a new version is available, only then will we know if this python/node-gyp integration works or whether it will still fallback to the availability of the prebuilt binaries from external sources. Only time will tell if we have solved the issue or not.